refactor(components/molecule/avatar): add import file extensions
