+++ components/molecule/avatar/demo/ArticleDefault.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import {Article, H2, Paragraph, Code} from '@s-ui/documentation-library'
 
-import MoleculeAvatar from '../src/index'
+import MoleculeAvatar from '../src/index.js'
 
 const ArticleDefault = ({className}) => {
   return (
