+++ components/molecule/avatar/demo/ArticleName.js
@@ -2,7 +2,7 @@ import {useState} from 'react'
 import PropTypes from 'prop-types'
 import {Article, H2, Input, Paragraph, Code} from '@s-ui/documentation-library'
 
-import MoleculeAvatar from '../src/index'
+import MoleculeAvatar from '../src/index.js'
 
 const ArticleName = ({className}) => {
   const [name, setName] = useState('John Doe')
