+++ components/molecule/avatar/src/AvatarBadge/index.js
@@ -1,26 +1,11 @@
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 
-export const AVATAR_BADGE_STATUSES = {
-  ERROR: 'error',
-  SUCCESS: 'success',
-  ALERT: 'alert',
-  MUTED: 'muted'
-}
-
-export const AVATAR_BADGE_PLACEMENTS = {
-  TOP: 'top',
-  BOTTOM: 'bottom'
-}
-
-export const AVATAR_BADGE_SIZES = {
-  XXLARGE: 'xxlarge',
-  XLARGE: 'xlarge',
-  LARGE: 'large',
-  MEDIUM: 'medium',
-  SMALL: 'small',
-  XSMALL: 'xsmall'
-}
+import {
+  AVATAR_BADGE_STATUSES,
+  AVATAR_BADGE_PLACEMENTS,
+  AVATAR_BADGE_SIZES
+} from './settings.js'
 
 const MoleculeAvatarBadge = ({
   className: classNameProp,
