+++ components/molecule/avatar/src/AvatarBadge/settings.js
@@ -1,0 +1,20 @@
+export const AVATAR_BADGE_STATUSES = {
+  ERROR: 'error',
+  SUCCESS: 'success',
+  ALERT: 'alert',
+  MUTED: 'muted'
+}
+
+export const AVATAR_BADGE_PLACEMENTS = {
+  TOP: 'top',
+  BOTTOM: 'bottom'
+}
+
+export const AVATAR_BADGE_SIZES = {
+  XXLARGE: 'xxlarge',
+  XLARGE: 'xlarge',
+  LARGE: 'large',
+  MEDIUM: 'medium',
+  SMALL: 'small',
+  XSMALL: 'xsmall'
+}
