+++ components/molecule/avatar/src/AvatarFallbackIcon/settings.js
@@ -1,0 +1,1 @@
+export const BASE_CLASS_NAME = 'sui-MoleculeAvatarFallbackIcon'
